code flow/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose
jinja2
python-multipart
